[SYCL] Fix SYCL `im2col` and `convert` Overflow with Large Dims (#9052)
* sycl: fix im2col overflow and sync with cuda


* sycl: fix convert overflow


* sycl: fix convert and dequantize


* sycl: fix ib in dmmv


* sycl:refine convert


* sycl: move downsample global_range into common


* test: add im2col and convert test cases


* test: make new cases only in sycl


* test: comment new test_cases for only local testing


---------

+++ ggml/src/ggml-sycl/dmmv.cpp
@@ -3,7 +3,8 @@
 #include "dequantize.hpp"
 #include "presets.hpp"
 
-static void convert_f16(const void * vx, const int ib, const int iqs, dfloat2 & v){
+
+static void convert_f16(const void * vx, const int64_t ib, const int iqs, dfloat2 & v){
     const sycl::half *x = (const sycl::half *)vx;
 
     // automatic half -> float type cast if dfloat == float
@@ -11,7 +12,7 @@ static void convert_f16(const void * vx, const int ib, const int iqs, dfloat2 & 
     v.y() = x[ib + iqs + 1];
 }
 
-static void convert_f32(const void * vx, const int ib, const int iqs, dfloat2 & v){
+static void convert_f32(const void * vx, const int64_t ib, const int iqs, dfloat2 & v){
     const float * x = (const float *) vx;
 
     // automatic half -> float type cast if dfloat == float
